requirements.txt
Django==1.9.4
django-crispy-forms==1.6.0
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
